requirements.txt
pandas=2.2.2
matplotlib==3.8.4
termcolor=2.4.0
tabulate==0.9.0
itermplot==0.331
tables==3.9.2
sqlalchemy==2.0.29
mysql-connector-python==8.3.0
